current test setup
