+++ tests/generic_full_path/more_schemas.rs
@@ -1,0 +1,4 @@
+use utoipa::ToSchema;
+
+#[derive(ToSchema)]
+pub struct MoreSchema;
